requirements.txt
Pillow==9.0.0
requests==2.22.0
tinydb==4.5.2
ttkbootstrap==1.2.0
